Make clang-tidy happy

+++ src/engine/image_tool.h
@@ -36,7 +36,7 @@ namespace fheroes2
 
     Sprite decodeICNSprite( const uint8_t * data, const uint32_t sizeData, const int32_t width, const int32_t height, const int16_t offsetX, const int16_t offsetY );
 
-    void decodeTILImages( const uint8_t * data, const int32_t imageCount, const int32_t width, const int32_t height, std::vector<Image> & output );
+    void decodeTILImages( const uint8_t * data, const size_t imageCount, const int32_t width, const int32_t height, std::vector<Image> & output );
 
     // By default only Bitmap (.bmp) format is supported.
     bool isPNGFormatSupported();
